+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/dto/request/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package Coordinate.coordikittyBE.domain.auth.login.dto.response;
+package Coordinate.coordikittyBE.domain.auth.login.dto.request;
 
 
 public record LoginRequestDto(String email, String password) {
